discord-agent/requirements.txt
discord.py
dotenv
pika
asyncio
requests
aiohttp
